requirements.txt
langchain-openai
langchain-core
langchain
python-dotenv
langchain_community
streamlit